exception en servicio-service

+++ servicio-service/src/main/java/com/barbershop/servicio_service/service/ServicioService.java
@@ -1,7 +1,9 @@
 package com.barbershop.servicio_service.service;
 
+import com.barbershop.servicio_service.exception.ResourceNotFoundException;
 import com.barbershop.servicio_service.model.Servicio;
 import com.barbershop.servicio_service.repository.ServicioRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,7 +20,11 @@ public class ServicioService {
     }
 
     public Servicio findById(Long id) {
-        return repository.findById(id).orElse(null);
+
+        return repository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException(
+                                "Servicio con ID " + id + " no encontrado"));
     }
 
     public Servicio save(Servicio servicio) {
@@ -26,6 +32,12 @@ public class ServicioService {
     }
 
     public void delete(Long id) {
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Servicio con ID " + id + " no encontrado");
+        }
+
         repository.deleteById(id);
     }
